moved generated js files to dist folder, add grunt task and dependency
