+++ package.json
@@ -41,6 +41,7 @@
     "bpmn-js": "^8.8.2",
     "diagram-js": "^7.8.1",
     "exceljs": "^4.3.0",
+    "grunt-shell": "^4.0.0",
     "jquery": "^3.6.0",
     "jszip": "^3.7.1"
   }
