--- planner/types/Dataclass.js
@@ -1,9 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.Dataclass = void 0;
-class Dataclass {
-    constructor(name) {
-        this.name = name;
-    }
-}
-exports.Dataclass = Dataclass;
