--- planner/types/Role.js
@@ -1,9 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.Role = void 0;
-class Role {
-    constructor(name) {
-        this.name = name;
-    }
-}
-exports.Role = Role;
