--- planner/types/executionState/DataObjectInstance.js
@@ -1,10 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.DataObjectInstance = void 0;
-class DataObjectInstance {
-    constructor(name, dataclass) {
-        this.name = name;
-        this.dataclass = dataclass;
-    }
-}
-exports.DataObjectInstance = DataObjectInstance;
