--- planner/types/executionState/InstanceLink.js
@@ -1,10 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.InstanceLink = void 0;
-class InstanceLink {
-    constructor(first, second) {
-        this.first = first;
-        this.second = second;
-    }
-}
-exports.InstanceLink = InstanceLink;
