--- planner/types/fragments/Activity.js
@@ -1,6 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.Activity = void 0;
-class Activity {
-}
-exports.Activity = Activity;
